+++ pkg/clientset/typed/apis/v1alpha1/generated_expansion.go
@@ -20,6 +20,8 @@ package v1alpha1
 
 type AIGatewayExpansion interface{}
 
+type DataPlaneKonnectExtensionExpansion interface{}
+
 type DataPlaneMetricsExtensionExpansion interface{}
 
 type KongPluginInstallationExpansion interface{}